NOKEtheMESSENGER/noke_messenger/noke_the_messenger/requirements.txt
flask==3.0.0
openpyxl==3.1.2
twilio==8.10.0
sendgrid==6.11.0
python-dotenv==1.0.0
